Restructure code to clean architecture
