--- src/main/java/com/ite/itea/application/dto/ItemRequest.java
@@ -1,7 +1,0 @@
-package com.ite.itea.application.dto;
-
-import com.ite.itea.domain.retail.ProductName;
-
-public record ItemRequest(ProductName name, int amount, long price) {
-
-}
