--- src/main/java/com/ite/itea/application/dto/OrderRequest.java
@@ -1,7 +1,0 @@
-package com.ite.itea.application.dto;
-
-import java.util.List;
-
-public record OrderRequest(List<ItemRequest> itemRequests) {
-
-}
