--- src/main/java/com/ite/itea/application/dto/Receipt.java
@@ -1,5 +1,0 @@
-package com.ite.itea.application.dto;
-
-public record Receipt(long priceInCents, String text) {
-
-}
