--- src/main/java/com/ite/itea/application/dto/ReceiptResponse.java
@@ -1,5 +1,0 @@
-package com.ite.itea.application.dto;
-
-public record ReceiptResponse(long priceInCents, String text) {
-
-}
