--- src/main/java/com/ite/itea/domain/retail/Chair.java
@@ -1,22 +1,0 @@
-package com.ite.itea.domain.retail;
-
-import com.ite.itea.domain.core.EuroPrice;
-
-public final class Chair extends Product {
-
-    private final EuroPrice legPrice;
-    private final EuroPrice seatPrice;
-    private final EuroPrice backRestPrice;
-
-    public Chair(ProductId id, String name, EuroPrice legPrice, EuroPrice seatPrice, EuroPrice backRestPrice) {
-        super(id, name);
-        this.legPrice = legPrice;
-        this.seatPrice = seatPrice;
-        this.backRestPrice = backRestPrice;
-    }
-
-    @Override
-    public EuroPrice price() {
-        return legPrice.times(4).plus(seatPrice).plus(backRestPrice);
-    }
-}
