--- src/main/java/com/ite/itea/domain/retail/Picture.java
@@ -1,18 +1,0 @@
-package com.ite.itea.domain.retail;
-
-import com.ite.itea.domain.core.EuroPrice;
-
-public final class Picture extends Product {
-
-    private final EuroPrice price;
-
-    public Picture(ProductId id, String name, EuroPrice price) {
-        super(id, name);
-        this.price = price;
-    }
-
-    @Override
-    public EuroPrice price() {
-        return price;
-    }
-}
