--- src/main/java/com/ite/itea/domain/retail/ProductRepository.java
@@ -1,8 +1,0 @@
-package com.ite.itea.domain.retail;
-
-import java.util.Optional;
-
-public interface ProductRepository {
-
-    Optional<Product> byId(ProductId id);
-}
