--- src/main/java/com/ite/itea/domain/retail/ShoppingCart.java
@@ -1,30 +1,0 @@
-package com.ite.itea.domain.retail;
-
-import com.ite.itea.domain.core.EuroPrice;
-
-import java.util.ArrayList;
-
-public final class ShoppingCart {
-
-    private final ArrayList<CartItem> cartItems = new ArrayList<>();
-
-    public ShoppingCart() {
-    }
-
-    public static ShoppingCart empty() {
-        return new ShoppingCart();
-    }
-
-    public void add(Product product) {
-        cartItems.add(new CartItem(product));
-    }
-
-    public EuroPrice price() {
-        return cartItems.stream()
-                .map(cartItem -> cartItem.product.price())
-                .reduce(EuroPrice.zero(), EuroPrice::plus);
-    }
-
-    private record CartItem(Product product) {
-    }
-}
