+++ src/main/java/com/ite/itea/retail/adapters/in/controller/ItemRequest.java
@@ -1,0 +1,7 @@
+package com.ite.itea.retail.adapters.in.controller;
+
+import com.ite.itea.retail.domain.entity.retail.ProductName;
+
+public record ItemRequest(ProductName name, int amount, long price) {
+
+}
