+++ src/main/java/com/ite/itea/retail/adapters/in/controller/OrderRequest.java
@@ -1,0 +1,7 @@
+package com.ite.itea.retail.adapters.in.controller;
+
+import java.util.List;
+
+public record OrderRequest(List<ItemRequest> itemRequests) {
+
+}
