+++ src/main/java/com/ite/itea/retail/adapters/in/controller/ReceiptResponse.java
@@ -1,0 +1,5 @@
+package com.ite.itea.retail.adapters.in.controller;
+
+public record ReceiptResponse(long priceInCents, String text) {
+
+}
