+++ src/main/java/com/ite/itea/retail/domain/entity/retail/ProductId.java
@@ -1,0 +1,10 @@
+package com.ite.itea.retail.domain.entity.retail;
+
+import java.util.UUID;
+
+public record ProductId(String internalID) {
+
+    public static ProductId random() {
+        return new ProductId(UUID.randomUUID().toString());
+    }
+}
