+++ src/main/java/com/ite/itea/retail/domain/usecase/ProductRepository.java
@@ -1,0 +1,11 @@
+package com.ite.itea.retail.domain.usecase;
+
+import com.ite.itea.retail.domain.entity.retail.Product;
+import com.ite.itea.retail.domain.entity.retail.ProductId;
+
+import java.util.Optional;
+
+public interface ProductRepository {
+
+    Optional<Product> byId(ProductId id);
+}
