+++ src/main/java/com/ite/itea/retail/domain/usecase/ReceiptPresenter.java
@@ -1,0 +1,8 @@
+package com.ite.itea.retail.domain.usecase;
+
+import com.ite.itea.retail.adapters.out.presenter.Receipt;
+import com.ite.itea.retail.domain.entity.retail.Order;
+
+public interface ReceiptPresenter {
+    Receipt prepareReceipt(Order order);
+}
